@@ -1,113 +1,85 @@
 {
-  "name": "gdl-frontend",
-  "version": "0.0.1",
-  "description": "GDL Frontend",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/GlobalDigitalLibraryio/gdl-frontend.git"
-  },
-  "bugs": {
-    "url": "https://github.com/GlobalDigitalLibraryio/gdl-frontend/issues"
-  },
-  "engines": {
-    "node": ">=6.6.0",
-    "npm": "~3.10.3"
-  },
-  "lingui": {
-    "fallbackLocale": "en",
-    "sourceLocale": "en",
-    "srcPathDirs": [
-      "<rootDir>/components/",
-      "<rootDir>/pages/",
-      "<rootDir>/hocs/",
-      "<rootDir>/elements/"
-    ]
-  },
-  "jest": {
-    "setupFiles": ["./jest.setup.js"]
-  },
-  "browser": {
-    "dns": false,
-    "raven": "raven-js"
+  "name": "gdl",
+  "private": true,
+  "bolt": {
+    "workspaces": ["packages/*"]
   },
   "author": "gdl@knowit.no",
+  "repository": "https://github.com/GlobalDigitalLibraryio/gdl-frontend.git",
   "license": "GPL-3.0",
   "scripts": {
+    "build": "bolt workspace gdl-frontend run build",
+    "build:frontend": "bolt workspace gdl-frontend run build",
+    "check": "flow",
     "test": "jest",
-    "lint": "eslint . --ext .js --max-warnings=0 --ignore-path .prettierignore",
-    "dev": "node server",
-    "build": "npm run i18n:extract && npm run i18n:compile && next build",
-    "start": "NODE_ENV=production node server",
-    "i18n:extract": "lingui extract --clean",
-    "i18n:compile": "lingui compile",
-    "i18n:add-locale": "lingui add-locale",
     "format": "prettier --write \"**/*.js\"",
     "format:check": "prettier --list-different \"**/*.js\"",
-    "flow": "flow",
-    "check":
-      "npm run lint && npm run format:check && npm run flow && npm run test"
-  },
-  "dependencies": {
-    "@lingui/react": "2.0.5",
-    "@zeit/next-source-maps": "^0.0.2",
-    "auth0-js": "9.4.2",
-    "cookie-parser": "1.4.3",
-    "domurl": "2.1.7",
-    "emotion": "9.1.2",
-    "emotion-server": "9.1.2",
-    "emotion-theming": "^9.1.2",
-    "express": "4.16.3",
-    "express-request-language": "1.1.12",
-    "glob": "7.1.2",
-    "helmet": "^3.12.0",
-    "invariant": "^2.2.4",
-    "isomorphic-unfetch": "2.0.0",
-    "jwt-decode": "^2.2.0",
-    "lru-cache": "^4.1.2",
-    "lscache": "1.1.0",
-    "next": "5.1.0",
-    "next-routes": "1.4.1",
-    "polished": "1.9.2",
-    "prop-types": "15.6.1",
-    "raven": "^2.5.0",
-    "raven-js": "^3.24.2",
-    "react": "16.3.2",
-    "react-dom": "16.3.2",
-    "react-emotion": "9.1.2",
-    "react-ga": "^2.4.0",
-    "react-icons": "2.2.7",
-    "react-json-view": "^1.16.1",
-    "react-portal": "^4.1.5",
-    "react-swipeable": "4.2.1",
-    "styled-map": "^2.0.1",
-    "styled-system": "2.2.1",
-    "styled-theming": "^2.2.0",
-    "universal-cookie": "^2.1.2"
+    "lint": "eslint . --ext .js --max-warnings=0 --ignore-path .prettierignore"
   },
   "devDependencies": {
-    "@sector-labs/lingui-babel-preset-react": "2.0.4-sl.6",
-    "@sector-labs/lingui-cli": "2.0.4-sl.6",
+    "@sector-labs/lingui-babel-preset-react": "^2.0.4-sl.6",
+    "@sector-labs/lingui-cli": "^2.0.4-sl.6",
     "babel-eslint": "^8.2.3",
-    "babel-plugin-emotion": "9.1.2",
-    "babel-plugin-inline-react-svg": "0.5.2",
-    "babel-plugin-polished": "1.1.0",
-    "babel-plugin-transform-assets": "0.2.0",
-    "babel-plugin-transform-flow-strip-types": "6.22.0",
-    "enzyme": "3.3.0",
-    "enzyme-adapter-react-16": "1.1.1",
-    "enzyme-to-json": "3.3.3",
-    "eslint": "4.19.1",
+    "babel-plugin-emotion": "^9.1.2",
+    "babel-plugin-inline-react-svg": "^0.5.2",
+    "babel-plugin-polished": "^1.1.0",
+    "babel-plugin-transform-assets": "^0.2.0",
+    "babel-plugin-transform-flow-strip-types": "^6.22.0",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "enzyme-to-json": "^3.3.4",
+    "eslint": "^4.19.1",
     "eslint-config-react-app": "^2.1.0",
-    "eslint-plugin-flowtype": "2.46.2",
-    "eslint-plugin-import": "2.11.0",
-    "eslint-plugin-jest": "^21.15.0",
-    "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-react": "7.7.0",
-    "flow-bin": "0.67.1",
-    "jest": "22.4.3",
-    "jest-emotion": "9.1.2",
-    "nock": "^9.2.5",
+    "eslint-plugin-flowtype": "^2.47.1",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-jest": "^21.15.2",
+    "eslint-plugin-jsx-a11y": "^5.1.0",
+    "eslint-plugin-react": "^7.8.2",
+    "flow-bin": "^0.67.1",
+    "jest": "^22.4.4",
+    "jest-emotion": "^9.1.3",
+    "nock": "^9.2.6",
     "prettier": "^1.12.1",
-    "react-test-renderer": "16.3.2"
+    "react-test-renderer": "^16.4.0"
+  },
+  "dependencies": {
+    "@lingui/react": "^2.0.8",
+    "@zeit/next-source-maps": "^0.0.2",
+    "auth0-js": "^9.5.1",
+    "cookie-parser": "^1.4.3",
+    "domurl": "^2.1.8",
+    "emotion": "^9.1.3",
+    "emotion-server": "^9.1.3",
+    "emotion-theming": "^9.1.2",
+    "express": "^4.16.3",
+    "express-request-language": "^1.1.15",
+    "glob": "^7.1.2",
+    "helmet": "^3.12.1",
+    "invariant": "^2.2.4",
+    "isomorphic-unfetch": "^2.0.0",
+    "jwt-decode": "^2.2.0",
+    "lru-cache": "^4.1.3",
+    "lscache": "^1.1.0",
+    "next": "^5.1.0",
+    "next-routes": "^1.4.2",
+    "polished": "^1.9.2",
+    "prop-types": "^15.6.1",
+    "raven": "^2.6.2",
+    "raven-js": "^3.25.2",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
+    "react-emotion": "^9.1.3",
+    "react-ga": "^2.5.2",
+    "react-icons": "^2.2.7",
+    "react-json-view": "^1.16.6",
+    "react-portal": "^4.1.5",
+    "react-swipeable": "^4.3.0",
+    "styled-map": "^2.0.1",
+    "styled-system": "^2.2.5",
+    "styled-theming": "^2.2.0",
+    "universal-cookie": "^2.1.5"
+  },
+  "engines": {
+    "node": ">=8.0.0"
   }
 }
